% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_beta_class.R
\name{add_beta_class}
\alias{add_beta_class}
\title{Add Beta Class to Input Data Frame}
\usage{
add_beta_class(x)
}
\arguments{
\item{x}{a data frame of daily stock return informtion, created by
\code{gather_daily()}.}
}
\value{
A data frame on stock-by-date information with a new \code{beta.class} column.
}
\description{
add_beta_class() takes an input data frame, calculates a measure of beta, adds that
measure to the input, and return the new data frame.
}

