% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_market_return.R
\name{calc_market_return}
\alias{calc_market_return}
\title{Calculate the Return of the Market}
\usage{
calc_market_return(x)
}
\arguments{
\item{x}{a data frame of daily stock return informtion, created by 
\code{clean_data()}.}
}
\value{
A data frame of daily market returns.
}
\description{
calc_market_return() takes an input data frame and calculates a measure of
the daily return of the market.
}

