% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Daily Data}
\usage{
clean_data()
}
\value{
A data frame on stock-by-date information.
}
\description{
clean_data() brings together all the daily return data we need from the 
ws.data package. This is not meant to be used interactively, since it
requires the ws.data package.
}

