% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_daily.R
\name{gather_daily}
\alias{gather_daily}
\title{Gather Daily Data}
\usage{
gather_daily()
}
\value{
A data frame on stock-by-date information.
}
\description{
gather_daily() brings together all the daily return data we need from the 
ws.data package. This is not meant to be used interactively, since it
requires the ws.data package.
}

