% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_monthly.R
\name{gather_monthly}
\alias{gather_monthly}
\title{Gather Monthly Data}
\usage{
gather_monthly(x)
}
\arguments{
\item{x}{a data frame of daily stock return informtion, created by
\code{gather_daily()}.}
}
\value{
A data frame on stock-by-date information on a monthly basis.
}
\description{
gather_monthly() takes a data frame for daily return data and returns a data 
frame of monthly data that is stored in data/ and used in the vignette. This 
is not meant to be used interactively.
}

